a function that searches for a value in a sorted array of integers using the Interpolation search algorithm

+++ 0x1E-search_algorithms/102-interpolation.c
@@ -28,7 +28,7 @@ int interpolation_search(int *array, size_t size, int value)
 		fraction = (double)(end - start) / (array[end]
 				- array[start]) * (value - array[start]);
 		position = (size_t)(start + fraction);
-		printf("Value checked array[%zu]", position);
+		printf("Value checked array[%d]", (int)position);
 
 		if (position >= size)
 		{
